Added random matrix generation
randomMatrix generates a random seed, which is then input into a set of
parameters to generate a unimodular matrix. However, since there are
only a limited number of possibilities before the numbers are too large,
this is not very random.

moreRandomMatrix generates a random unimodular matrix by randomly
generating one from randomMatrix, then randomly inverts it and
multiplies it by a random number of unimodular matrices.